Tests + Fixed Pixel Stuff
Fixes the GetPixel() and SetPixel() offset calculation...again.
Apparently, we were supposed to use the Stride of the image, not the
width, which seems odd to that it actually still worked before.

Probably means that we have to rerun our database fingerprinting
algorithm, but we can write a utility that checks to see if that is
necessary.